Type mismatch bug

+++ spd/harvest/harvest.py
@@ -205,7 +205,7 @@ def harvest_activation_contexts(
     logger.info(f"Loading model on {device}")
 
     run_info = SPDRunInfo.from_path(config.wandb_path)
-    model = ComponentModel.from_run_info(run_info).to(device)
+    model = ComponentModel.from_run_info(run_info).float().to(device)
     model.eval()
 
     spd_config = run_info.config
